% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgd.R
\name{cdgd}
\alias{cdgd}
\title{Decomposition based on conditional ignorability}
\usage{
cdgd(
  Y,
  D,
  G1,
  G2,
  Q,
  X,
  data,
  weight = NULL,
  alpha = 0.05,
  k = 500,
  t = 0.05,
  algorithm
)
}
\arguments{
\item{Y}{Outcome. The name of a continuous variable.}

\item{D}{Treatment status. The name of a binary numeric variable taking values of 0 and 1.}

\item{G1}{Group 1 membership. The name of a binary factor variable taking values of 0 and 1.}

\item{G2}{Group 2 membership. The name of a binary factor variable taking values of 0 and 1.}

\item{Q}{Conditional set. The vector of the names of numeric variables.}

\item{X}{Confounders. The vector of the names of numeric variables.}

\item{data}{A data frame.}

\item{weight}{Survey weights. The name of a numeric variable.}

\item{alpha}{(1-alpha) confidence intervals.}

\item{k}{Number of monte carlo simulation.}

\item{t}{Threshold of propensity score censoring. Propensity scores larger than 1-t or smaller than t will be censored.}

\item{algorithm}{The ML algorithm for modelling. "nnet" for neural network and "ranger" for random forests.}
}
\value{
A data frame of point estimates and confidence intervals
}
\description{
Decomposition based on conditional ignorability
}
\examples{
set.seed(1)
results <- cdgd(Y="outcome",D="treatment",G1="group_a",G2="group_b",X=c("confounder","Q"),
Q="Q",data=exp_data,t=0.05,algorithm="nnet",alpha=0.05,k=20)
results
}
