% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgd0.R
\name{cdgd0}
\alias{cdgd0}
\title{Only produce point estimates of cdgd}
\usage{
cdgd0(Y, D, G1, G2, Q = NULL, X, data, weight = NULL, t = 0.05, algorithm)
}
\arguments{
\item{Y}{Outcome. The name of a continuous variable.}

\item{D}{Treatment status. The name of a binary numeric variable taking values of 0 and 1.}

\item{G1}{Group 1 membership. The name of a binary factor variable taking values of 0 and 1.}

\item{G2}{Group 2 membership. The name of a binary factor variable taking values of 0 and 1.}

\item{Q}{Conditional set. The vector of the names of numeric variables.}

\item{X}{Confounders. The vector of the names of numeric variables.}

\item{data}{A data frame.}

\item{weight}{Survey weights. The name of a numeric variable.}

\item{t}{Threshold of propensity score censoring. Propensity scores larger than 1-t or smaller than t will be censored.}

\item{algorithm}{The ML algorithm for modelling. "nnet" for neural network and "ranger" for random forests.}
}
\value{
A data frame of point estimates.
}
\description{
Only produce point estimates of cdgd
}
\examples{
data(exp_data)

set.seed(1)

results0 <- cdgd0(
Y="outcome",
D="treatment",
G1="group_a",
G2="group_b",
X=c("confounder","Q"),
Q="Q",
data=exp_data,
t=0.05,
algorithm="nnet")

results0
}
