% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgd1_ml.R
\name{cdgd1_ml}
\alias{cdgd1_ml}
\title{Perform conditional decomposition via Machine Learning}
\usage{
cdgd1_ml(Y, D, G, X, Q, data, algorithm, alpha = 0.05)
}
\arguments{
\item{Y}{Outcome. The name of a continuous variable.}

\item{D}{Treatment status. The name of a binary numeric variable taking values of 0 and 1.}

\item{G}{Advantaged group membership. The name of a binary numeric variable taking values of 0 and 1.}

\item{X}{Confounders. The vector of the names of numeric variables.}

\item{Q}{Conditional set. The vector of the names of numeric variables.}

\item{data}{A data frame.}

\item{algorithm}{The ML algorithm for modelling. "nnet" for neural network, "ranger" for random forests, "gbm" for generalized boosted models.}

\item{alpha}{1-alpha confidence interval.}
}
\value{
A dataframe of estimates.
}
\description{
Perform conditional decomposition via Machine Learning
}
\examples{
data(exp_data)

set.seed(1)

results <- cdgd1_ml(
Y="outcome",
D="treatment",
G="group_a",
X="confounder",
Q="Q",
data=exp_data,
algorithm="gbm")

results
}
