% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgd1_pa.R
\name{cdgd1_pa}
\alias{cdgd1_pa}
\title{Perform conditional decomposition via parametric models}
\usage{
cdgd1_pa(Y, D, G, X, Q, data, algorithm, alpha = 0.05)
}
\arguments{
\item{Y}{Outcome. The name of a continuous variable.}

\item{D}{Treatment status. The name of a binary numeric variable taking values of 0 and 1.}

\item{G}{Advantaged group membership. The name of a binary numeric variable taking values of 0 and 1.}

\item{X}{Confounders. The vector of the names of numeric variables.}

\item{Q}{Conditional set. The vector of the names of numeric variables.}

\item{data}{A data frame.}

\item{algorithm}{The ML algorithm for modelling. "nnet" for neural network, "ranger" for random forests, "gbm" for generalized boosted models.}

\item{alpha}{1-alpha confidence interval.}
}
\value{
A data frame of point estimates.
}
\description{
Perform conditional decomposition via parametric models
}
\examples{
data(exp_data)

results0 <- cdgd0(
Y="outcome",
D="treatment",
G="group_a",
#' X=c("confounder","Q"),
Q="Q",
data=exp_data,
t=0.05,
algorithm="nnet")

results0
}
