% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_data.R
\docType{data}
\name{exp_data}
\alias{exp_data}
\title{Simulated example data}
\format{
An object of class \code{data.frame} with 1000 rows and 5 columns.
}
\usage{
data(exp_data)
}
\description{
Simulated example data
}
\keyword{datasets}
