% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgd.R
\name{cdgd}
\alias{cdgd}
\title{Decomposition based on conditional ignorability}
\usage{
cdgd(
  Y,
  W,
  G1,
  G2,
  Q,
  X,
  data,
  alpha = 0.05,
  weight = NULL,
  k = 500,
  t = 0.05,
  algorithm
)
}
\arguments{
\item{Y}{The name of a continuous variable.}

\item{W}{The name of a binary factor variable.}
}
\value{
A list of two vectors. The first is the point estimates, the second is standard errors.
}
\description{
Decomposition based on conditional ignorability
}
