% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgd0_manual.R
\name{cdgd0_manual}
\alias{cdgd0_manual}
\title{Perform unconditional decomposition using conditional expectations predicted beforehand.}
\usage{
cdgd0_manual(
  Y,
  D,
  G,
  YgivenX.Pred_D0G0,
  YgivenX.Pred_D1G0,
  YgivenX.Pred_D0G1,
  YgivenX.Pred_D1G1,
  DgivenX.Pred_G0,
  DgivenX.Pred_G1,
  data,
  alpha = 0.05
)
}
\arguments{
\item{Y}{Outcome. The name of a continuous variable.}

\item{D}{Treatment status. The name of a binary numeric variable taking values of 0 and 1.}

\item{G}{Advantaged group membership. The name of a binary numeric variable taking values of 0 and 1.}

\item{YgivenX.Pred_D0G0}{A numeric vector of predicted Y values given X, D=0, and G=0. Vector length=nrow(data).}

\item{YgivenX.Pred_D1G0}{A numeric vector of predicted Y values given X, D=1, and G=0. Vector length=nrow(data).}

\item{YgivenX.Pred_D0G1}{A numeric vector of predicted Y values given X, D=0, and G=1. Vector length=nrow(data).}

\item{YgivenX.Pred_D1G1}{A numeric vector of predicted Y values given X, D=1, and G=1. Vector length=nrow(data).}

\item{DgivenX.Pred_G0}{A numeric vector of predicted D values given X and G=1. Vector length=nrow(data).}

\item{DgivenX.Pred_G1}{A numeric vector of predicted D values given X and G=0. Vector length=nrow(data).}

\item{data}{A data frame.}

\item{alpha}{1-alpha confidence interval.}
}
\value{
A list of estimates.
}
\description{
The conditional expectations, i.e., the nuisance terms, should be estimated using cross-fitting.
}
